requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
psycopg[binary]==3.2.9
psycopg_pool==3.2.7
python-dotenv==1.1.1
python-jose[cryptography]==3.5.0
bcrypt==4.1.3
pydantic==2.11.3
email-validator==2.2.0
python-multipart==0.0.20
pillow==11.2.1
sqlalchemy==2.0.40
sqladmin==0.20.1
jinja2==3.1.6
alembic==1.16.4
redis==5.2.1
prometheus-client==0.22.1
anyio==4.9.0
itsdangerous==2.2.0
boto3==1.37.38
timm==1.0.15
torch>=2.5.0,<3.0
